Done with Controller, Service and Repository TDD.

+++ src/main/java/com/bazaar/todo/dto/TodoListItemDto.java
@@ -1,19 +1,19 @@
 package com.bazaar.todo.dto;
 
 public class TodoListItemDto {
-    private int id;
+    private Long id;
     private String title;
 
-    public TodoListItemDto(int id, String title) {
+    public TodoListItemDto(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
